Portfolio evaluation 1. Implemented: periodically started service that calculates portfolio valuation at the end of the previous day. 2. Added to settings fragment: Selector at what time user wants (if so) to receive daily notification about portfolio state. 3. SharedPreferences were switched to Kotlin delegates scheme. 4. InfoFragment tags & links switched to FlexBox, icons added to known links types. 5. Link Stat added

+++ coin_paprika_api/src/main/java/dev/kokorev/coin_paprika_api/entity/LinkExtendedEntity.kt
@@ -4,5 +4,5 @@ package com.coinpaprika.apiclient.entity
 data class LinkExtendedEntity(
     val url: String,
     val type: String,
-    val stats: Map<String, Int>?
+    val stats: Map<String, Long>?
 )
